Fixing too many wallet start retries
