+++ federator/src/lib/HathorWallet.ts
@@ -1,233 +1,72 @@
-// import axiosCurlirize from 'axios-curlirize';
 import axios from 'axios';
-import { LogWrapper } from './logWrapper';
-import { ConfigData } from './config';
-import { PubSub } from '@google-cloud/pubsub';
-import { HathorTx } from '../types/HathorTx';
-import { HathorUtxo } from '../types/HathorUtxo';
-import { HathorException } from '../types/HathorException';
-import { BridgeFactory } from '../contracts/BridgeFactory';
-import { FederationFactory } from '../contracts/FederationFactory';
-import { EvmBroker } from './Broker/EvmBroker';
-import { Broker } from './Broker/Broker';
+import EventEmmiter from 'node:events';
+// import axiosCurlirize from 'axios-curlirize';
 import { HathorResponse, StatusResponse } from '../types/HathorResponseTypes';
+import { LogWrapper } from './logWrapper';
 import { ConfigChain } from './configChain';
-import { HathorBroker } from './Broker/HathorBroker';
+import { ConfigData } from './config';
 
-// axiosCurlirize(axios);
+// axiosCurlirize(axios); // <- just for testing
+
+export type Wallet = {
+  ready: boolean;
+  lastCheck: Date;
+};
 
 export class HathorWallet {
-  public logger: LogWrapper;
-  public config: ConfigData;
-  public bridgeFactory: BridgeFactory;
-  public federationFactory: FederationFactory;
-  protected chainConfig: ConfigChain;
+  private static wallet: HathorWallet;
 
   private WALLET_STATUS_CONNECTING = 1;
   private WALLET_STATUS_SYNCING = 2;
   private WALLET_STATUS_READY = 3;
-  private nonRetriableErrors = [
-    /Invalid transaction. At least one of your inputs has already been spent./,
-    /Transaction already exists/,
-    /Invalid tx/,
-  ];
 
-  constructor(
-    config: ConfigData,
-    logger: LogWrapper,
-    bridgeFactory: BridgeFactory,
-    federationFactory: FederationFactory,
-  ) {
-    this.config = config;
+  public logger: LogWrapper;
+  public chainConfig: ConfigChain;
+  private wallets: Map<string, Wallet>;
+  public walletEmmiter = new EventEmmiter();
+
+  private constructor(config: ConfigData, logger: LogWrapper) {
     this.logger = logger;
     this.chainConfig = config.sidechain[0];
-    this.bridgeFactory = bridgeFactory;
-    this.federationFactory = federationFactory;
+    this.wallets = new Map<string, Wallet>();
+    this.wallets.set('multisig', { ready: false, lastCheck: new Date(0) });
+    this.wallets.set('single', { ready: false, lastCheck: new Date(0) });
   }
 
-  async sendTokensToHathor(receiverAddress: string, qtd: string, tokenAddress: string, txHash: string) {
-    if (this.chainConfig.multisigOrder > 1) return;
-
-    await this.isWalletReady(true);
-    await this.isWalletReady(false);
-
-    const broker = new EvmBroker(this.config, this.logger, this.bridgeFactory, this.federationFactory);
-    await broker.sendTokensToHathor(receiverAddress, qtd, tokenAddress, txHash);
+  public static getInstance(config: ConfigData, logger: LogWrapper): HathorWallet {
+    if (!HathorWallet.wallet) {
+      HathorWallet.wallet = new HathorWallet(config, logger);
+    }
+    return HathorWallet.wallet;
   }
 
-  async listenToEventQueue(): Promise<void> {
-    await this.isWalletReady(true);
-    switch (this.chainConfig.eventQueueType) {
-      case 'pubsub':
-        this.listenToPubSubEventQueue();
-        break;
-      case 'sqs':
-        this.logger.error('AWS SQS not implemented.');
-        break;
-      case 'asb':
-        this.logger.error('Azure Service Bus not implemented.');
-        break;
+  public async areWalletsReady(): Promise<[boolean, EventEmmiter]> {
+    const multisig = this.wallets.get('multisig');
+    const single = this.wallets.get('single');
+    const currentTime = new Date();
+    const oneHourAgo = currentTime.getTime() - 10 * 60 * 1000;
+
+    if (multisig.lastCheck.getTime() < oneHourAgo || single.lastCheck.getTime() < oneHourAgo) {
+      this.isReady(true);
+      this.isReady(false);
+      return [false, this.walletEmmiter];
+    }
+
+    if (multisig.ready && single.ready) {
+      return [true, null];
+    }
+
+    return [false, this.walletEmmiter];
+  }
+
+  private setWalletReady(wallet: string) {
+    this.wallets.set(wallet, { ready: true, lastCheck: new Date() });
+    if (this.wallets.get('multisig').ready && this.wallets.get('single').ready) {
+      this.walletEmmiter.emit('wallets-ready');
     }
   }
 
-  private async listenToPubSubEventQueue() {
-    const pubsub = new PubSub({ projectId: this.chainConfig.pubsubProjectId });
-
-    const [subscriptions] = await pubsub.getSubscriptions();
-    const subscription = subscriptions.find(
-      (sub) =>
-        sub.name ===
-        `projects/${this.chainConfig.pubsubProjectId}/subscriptions/hathor-federator-${this.chainConfig.multisigOrder}-sub`,
-    );
-
-    if (subscription) {
-      subscription.on('message', async (message) => {
-        try {
-          const response = await this.parseHathorLogs(JSON.parse(message.data.toString()));
-          if (response) {
-            message.ack();
-            return;
-          }
-          message.nack();
-          return;
-        } catch (error) {
-          // TODO retry policy if fails to parse and resolve
-          this.logger.error(`Fail to processing hathor event: ${error}`);
-          let isNonRetriable = false;
-          if (error instanceof HathorException) {
-            const originalError = (error as HathorException).getOriginalMessage();
-
-            for (let index = 0; index < this.nonRetriableErrors.length; index++) {
-              const rgxError = this.nonRetriableErrors[index];
-              if (originalError.match(rgxError)) {
-                isNonRetriable = true;
-                break;
-              }
-            }
-
-            if (isNonRetriable) {
-              message.ack();
-              return;
-            }
-          }
-
-          message.nack();
-        }
-      });
-      subscription.on('error', (error) => {
-        this.logger.error(
-          `Unable to subscribe to topic hathor-federator-${this.chainConfig.multisigOrder}. Make sure it exists. Error: ${error}`,
-        );
-      });
-    } else {
-      this.logger.error(
-        `Unable to subscribe to topic hathor-federator-${this.chainConfig.multisigOrder}. Make sure it exists.`,
-      );
-    }
-  }
-
-  private async parseHathorLogs(event: any): Promise<boolean> {
-    if (event.type !== 'wallet:new-tx') return true;
-
-    await this.isWalletReady(true);
-    await this.isWalletReady(false);
-
-    const outUtxos = event.data.outputs.map(
-      (o) =>
-        new HathorUtxo(o.script, o.token, o.value, {
-          type: o.decoded?.type,
-          address: o.decoded?.address,
-          timelock: o.decoded?.timelock,
-        }),
-    );
-
-    const inUtxos = event.data.inputs.map(
-      (o) =>
-        new HathorUtxo(o.script, o.token, o.value, {
-          type: o.decoded?.type,
-          address: o.decoded?.address,
-          timelock: o.decoded?.timelock,
-        }),
-    );
-    const tx = new HathorTx(event.data.tx_id, event.data.timestamp, outUtxos, inUtxos);
-
-    const isProposal = tx.haveCustomData('hex');
-
-    if (isProposal) {
-      this.logger.info('Evaluating proposal...');
-      const txHex = tx.getCustomData('hex');
-      const [broker, dataType] = this.getBrokerAndDataType(tx);
-      const confirmed = await broker.isTxConfirmed(tx.tx_id);
-      if (!confirmed) {
-        return false;
-      }
-      await broker.sendMySignaturesToProposal(txHex, tx.getCustomData(dataType));
-      return true;
-    }
-
-    const isSignature = tx.haveCustomData('sig');
-
-    if (isSignature && this.chainConfig.multisigOrder >= this.chainConfig.multisigRequiredSignatures) {
-      this.logger.info('Evaluating signature...');
-      const [broker, dataType] = this.getBrokerAndDataType(tx);
-
-      const confirmed = await broker.isTxConfirmed(tx.tx_id);
-      if (!confirmed) {
-        return false;
-      }
-
-      const txId = tx.getCustomData(dataType);
-      const components = await broker.getSignaturesToPush(txId);
-
-      if (components.signatures.length < this.chainConfig.multisigRequiredSignatures) {
-        this.logger.info(
-          `Number of signatures not reached. Require ${this.chainConfig.multisigRequiredSignatures} signatures, have ${components.signatures.length}`,
-        );
-        return false;
-      }
-
-      await broker.signAndPushProposal(components.hex, txId, components.signatures);
-      return true;
-    }
-
-    const isHathorToEvm = tx.haveCustomData('addr');
-
-    if (isHathorToEvm) {
-      const broker = new HathorBroker(this.config, this.logger, this.bridgeFactory, this.federationFactory);
-
-      const confirmed = await broker.isTxConfirmed(tx.tx_id);
-      if (!confirmed) {
-        return false;
-      }
-
-      const tokenData = tx.getCustomTokenData()[0];
-
-      const result = await broker.sendTokensFromHathor(
-        tx.getCustomData('addr'),
-        tokenData.value,
-        tokenData.token,
-        tx.tx_id,
-        tokenData.sender,
-        tx.timestamp,
-      );
-      if (!result) {
-        throw new HathorException('Invalid tx', 'Invalid tx'); //TODO change exception type
-      }
-      return true;
-    }
-  }
-
-  private getBrokerAndDataType(tx: HathorTx): [Broker, string] {
-    const customDataType = tx.getCustomDataType();
-    switch (customDataType) {
-      case 'hsh':
-        return [new EvmBroker(this.config, this.logger, this.bridgeFactory, this.federationFactory), customDataType];
-      case 'hid':
-        return [new HathorBroker(this.config, this.logger, this.bridgeFactory, this.federationFactory), customDataType];
-    }
-  }
-
-  protected async isWalletReady(multisig: boolean, retry = 1): Promise<boolean> {
+  private async isReady(multisig: boolean, retry = 1): Promise<boolean> {
     const id = multisig ? this.chainConfig.multisigWalletId : this.chainConfig.singleWalletId;
     if (retry > 3) {
       this.logger.error(`Fail to start ${id} wallet: Maximum number of retries reached.`);
@@ -246,25 +85,26 @@ export class HathorWallet {
       const response = await axios.get<StatusResponse>(url, config);
       if (response.data.statusCode === this.WALLET_STATUS_READY) {
         this.logger.info(`${id} wallet is ready.`);
+        this.setWalletReady(multisig ? 'multisig' : 'single');
         return true;
       }
       if ([this.WALLET_STATUS_CONNECTING, this.WALLET_STATUS_SYNCING].includes(response.data.statusCode)) {
-        this.logger.info(`${id} wallet is ${response.data.statusMessage}.`);
-        await this.delay(3000);
-        return this.isWalletReady(multisig, ++retry);
+        this.logger.info(`${id} wallet is ${response.data.statusMessage ?? response.data.message}.`);
+        await this.delay(10000);
+        return this.isReady(multisig, ++retry);
       }
       if (!response.data.success && response.data.statusMessage === '') {
         this.logger.info(`${id} wallet looks stopped.`);
-        await this.startWallet(multisig);
-        await this.delay(3000);
-        return this.isWalletReady(multisig, ++retry);
+        await this.start(multisig);
+        await this.delay(10000);
+        return this.isReady(multisig, ++retry);
       }
     } catch (error) {
       throw Error(`Fail to get status of ${id} wallet: ${error}`);
     }
   }
 
-  private async startWallet(multisig: boolean): Promise<boolean> {
+  private async start(multisig: boolean): Promise<boolean> {
     const id = multisig ? this.chainConfig.multisigWalletId : this.chainConfig.singleWalletId;
     const seedKey = multisig ? this.chainConfig.multisigSeedKey : this.chainConfig.singleSeedKey;
     this.logger.info(`Trying to start ${id} wallet.`);
@@ -292,5 +132,3 @@ export class HathorWallet {
     return new Promise((resolve) => setTimeout(resolve, time));
   }
 }
-
-export default HathorWallet;
