+++ federator/src/types/HathorResponseTypes.ts
@@ -10,5 +10,5 @@ export type CreateProposalResponse = HathorResponse & { txHex: string };
 export type GetAddressResponse = HathorResponse & { address: string };
 export type GetConfirmationResponse = HathorResponse & { confirmationNumber: number };
 export type GetMySignatureResponse = HathorResponse & { signatures: string };
-export type StatusResponse = HathorResponse & { statusCode: number; statusMessage: string };
+export type StatusResponse = HathorResponse & { statusCode: number; statusMessage: string; message: string };
 export type DecodeResponse = HathorResponse & { tx: Data };
